fix: refactor search results logic
